1、修改更新用户角色异常提示； 2、添加角色同时绑定菜单； 3、修改角色同时绑定菜单； --待测试

+++ ddshteam-web-system-service/src/main/java/com/ddshteam/web/system/service/impl/SysUserServiceImpl.java
@@ -94,13 +94,17 @@ public class SysUserServiceImpl implements SysUserService {
 	public boolean updateUser(SysUserInfo SysUserInfo, String... roleIds) {
 		int result = sysUserInfoInfoDao.updateByPrimaryKeySelective(SysUserInfo);
 
-		SysRoleToUserCriteria sysRoleToUserCriteria = new SysRoleToUserCriteria();
-		com.ddshteam.web.system.service.api.model.SysRoleToUserCriteria.Criteria criteria = sysRoleToUserCriteria
-				.createCriteria();
-		criteria.andUserIdEqualTo(SysUserInfo.getId());
-		result = sysRoleToUserDao.deleteByExample(sysRoleToUserCriteria);
+		if(result>0&&roleIds!=null&&roleIds.length>0)
+		{
+			SysRoleToUserCriteria sysRoleToUserCriteria = new SysRoleToUserCriteria();
+			com.ddshteam.web.system.service.api.model.SysRoleToUserCriteria.Criteria criteria = sysRoleToUserCriteria
+					.createCriteria();
+			criteria.andUserIdEqualTo(SysUserInfo.getId());
+			sysRoleToUserDao.deleteByExample(sysRoleToUserCriteria);
 
-		result = sysRoleToUserCustomizeDao.insertRoleToUsers(SysUserInfo.getId(), roleIds);
+			sysRoleToUserCustomizeDao.insertRoleToUsers(SysUserInfo.getId(), roleIds);
+		}
+
 		// SysUserInfoInfoCustomizeDao.updateUser(SysUserInfo, roleIds);
 		return result > 0;
 	}
